Add ducks pattern
